Add on_failure_to_email and on_pass_to_email config keys
These keys can be used to override where pass/fail emails go. If they
are defined, emails are sent to the specified email. If not, they go
to the commit author.

This feature is sponsored by IRC insultbots and the letters B, A, C,
E and K.

No bits where harmed in the making of this feature.